backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
requests
python-multipart
python-dotenv
sqlalchemy
